[FIX] Smooth scrolling now works on page change.
Bug occurred due to DOM not being loaded when the document
query was resolved. We use setTimeout to wait for one render frame
and the DOM is ready to get the element with the fragment as Id.
Also adds a single centralized Scroll events handler to avoid having
too many subscribers on the router events.
With this new handler, scrolling smoothly on provided position and
on top of the page after routing location change is implemented.
