+++ src/app/core/routes/dynamic-routes.ts
@@ -1,4 +1,4 @@
-import { Routes, Route, Router } from '@angular/router';
+import { Routes, Router } from '@angular/router';
 
 export function registerDynamicRoutes(router: Router, getDynamicRoutes: () => Routes) {
     router.resetConfig([ ...getDynamicRoutes(), ...router.config ]);
